upd: lints and checks, move model dir in internal/app
