+++ internal/app/apiserver/server_internal_test.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"github.com/gorilla/securecookie"
 	"github.com/gorilla/sessions"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store/teststore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
