+++ internal/app/model/emailschedule.go
