+++ internal/app/model/emailschedule_test.go
@@ -1,7 +1,7 @@
 package model_test
 
 import (
-	"github.com/nizepart/rest-go/model"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
