+++ internal/app/model/message.go
