+++ internal/app/model/message_test.go
@@ -2,7 +2,7 @@ package model_test
 
 import (
 	"fmt"
-	"github.com/nizepart/rest-go/model"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
