+++ internal/app/store/sqlstore/emailtmpltrepo_test.go
@@ -1,8 +1,8 @@
 package sqlstore_test
 
 import (
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store/sqlstore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
