+++ internal/app/store/sqlstore/schedulerepo.go
@@ -1,7 +1,7 @@
 package sqlstore
 
 import (
-	"github.com/nizepart/rest-go/model"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"strconv"
 	"time"
 )
