+++ internal/app/store/sqlstore/schedulerepo_test.go
@@ -2,8 +2,8 @@ package sqlstore_test
 
 import (
 	"github.com/nizepart/rest-go/internal/app"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store/sqlstore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
