+++ internal/app/store/sqlstore/usrrepo_test.go
@@ -1,9 +1,9 @@
 package sqlstore_test
 
 import (
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store"
 	"github.com/nizepart/rest-go/internal/app/store/sqlstore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
