+++ internal/app/store/teststore/emailtmpltrepo.go
@@ -1,8 +1,8 @@
 package teststore
 
 import (
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store"
-	"github.com/nizepart/rest-go/model"
 )
 
 type EmailTemplateRepository struct {
