+++ internal/app/store/teststore/emailtmpltrepo_test.go
@@ -1,8 +1,8 @@
 package teststore_test
 
 import (
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store/teststore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
