+++ internal/app/store/teststore/schedulerepo_test.go
@@ -2,8 +2,8 @@ package teststore_test
 
 import (
 	"github.com/nizepart/rest-go/internal/app"
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store/teststore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
