+++ internal/app/store/teststore/userrepo_test.go
@@ -1,9 +1,9 @@
 package teststore_test
 
 import (
+	"github.com/nizepart/rest-go/internal/app/model"
 	"github.com/nizepart/rest-go/internal/app/store"
 	"github.com/nizepart/rest-go/internal/app/store/teststore"
-	"github.com/nizepart/rest-go/model"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
